requirements.txt
streamlit==1.38.0
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
protobuf==3.20.3
click==8.1.7